convert elapsed time factor to float
unsigned long lacks decimal

+++ Test Code/Accelerometer/Accelerometer.ino
@@ -14,7 +14,7 @@ float position[3] = {0, 0, 0};
 
 //initialize time keeping variables
 unsigned long loop_start_time;
-unsigned long elapsed_time_factor;
+float elapsed_time_factor;
 
 void setup() {
   Serial.begin(9600);
